@@ -1,7 +1,11 @@
 <script lang="ts">
-	import * as m from '$lib/paraglide/messages';
+	import type { Session } from '$lib/auth';
 	import LanguageSwitch from '$lib/components/language-switch.svelte';
 	import ThemeSwitch from '$lib/components/theme-switch.svelte';
+	import * as m from '$lib/paraglide/messages';
+	import NavAvater from './home-avater.svelte';
+
+	let { session }: { session: Session } = $props();
 </script>
 
 <header class="px-4 lg:px-6 h-14 flex items-center">
@@ -15,5 +19,6 @@
 		<a class="text-sm font-medium hover:underline underline-offset-4" href="/"> Contact </a>
 		<LanguageSwitch />
 		<ThemeSwitch />
+		<NavAvater {session} />
 	</nav>
 </header>
